refactor target checking, move out of context.rs and rename MaybeWarn to Policy
