+++ compiler/rustc_attr_parsing/src/attributes/stability.rs
@@ -12,10 +12,11 @@ use rustc_span::{Ident, Span, Symbol, sym};
 use super::util::parse_version;
 use super::{AcceptMapping, AttributeParser, OnDuplicate};
 use crate::attributes::NoArgsAttributeParser;
-use crate::context::MaybeWarn::Allow;
-use crate::context::{AcceptContext, AllowedTargets, FinalizeContext, Stage};
+use crate::context::{AcceptContext, FinalizeContext, Stage};
 use crate::parser::{ArgParser, MetaItemParser};
 use crate::session_diagnostics::{self, UnsupportedLiteralReason};
+use crate::target_checking::AllowedTargets;
+use crate::target_checking::Policy::Allow;
 
 macro_rules! reject_outside_std {
     ($cx: ident) => {
